app/account/views.py
from django.contrib.auth import get_user_model

from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status

from rest_framework_simplejwt.tokens import RefreshToken

from .serializer import RegistrationSerializer

User = get_user_model()


class RegisterUserView(APIView):
    def post(self, request):
        serializer = RegistrationSerializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        user = serializer.save()

        refresh = RefreshToken.for_user(user)
        response_data = {
            'token': str(refresh.access_token),
            'refreshToken': str(refresh),
        }
        return Response(response_data, status=status.HTTP_201_CREATED)
